% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_branch_contributions.R
\name{add_axis}
\alias{add_axis}
\title{Add axis to the color bar}
\usage{
add_axis(side = 4, cex.axis = 1.5, ...)
}
\arguments{
\item{side}{See \link[graphics:axis]{graphics::axis}. Default to \code{4}.}

\item{cex.axis}{See \link[graphics:axis]{graphics::axis}. Default to \code{1.5}.}

\item{...}{Additional parameters passed to \link[graphics:axis]{graphics::axis}.}
}
\value{
No value, but produces the side effect of adding an axis to the
color bar.
}
\description{
\code{add_axis()} adds an axis to the color bar.
}
\keyword{internal}
