% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_branch_contributions.R
\name{add_colorbar}
\alias{add_colorbar}
\title{Add color bar to the plot}
\usage{
add_colorbar(
  color_breaks,
  color_palette,
  dim,
  text_args = list(),
  axis_args = list(),
  ...
)
}
\arguments{
\item{color_breaks}{A string vector representing the boundaries of value
intervals so that values falling in the same interval have the same color.}

\item{color_palette}{A string vector representing the color palette for the
color bar.}

\item{dim}{An integer representing the dimension of eigenvector for plotting.}

\item{text_args}{A named list of arguments passed to an internal helper that
adds title text to the color bar. Elements include:
\describe{
\item{side}{See \link[graphics:mtext]{graphics::mtext}. Default to \code{4}.}
\item{line}{See \link[graphics:mtext]{graphics::mtext}. Default to \code{4}.}
\item{cex}{See \link[graphics:mtext]{graphics::mtext}. Default to \code{1.5}.}
\item{...}{Additional parameters passed to \link[graphics:mtext]{graphics::mtext}.}
}}

\item{axis_args}{A named list of arguments passed to an internal helper that
adds axis to the color bar.
\describe{
\item{side}{See \link[graphics:axis]{graphics::axis}. Default to \code{4}.}
\item{cex.axis}{See \link[graphics:axis]{graphics::axis}. Default to \code{1.5}.}
\item{...}{Additional parameters passed to \link[graphics:axis]{graphics::axis}.}
}}

\item{...}{Parameters passed to \link[graphics:image]{graphics::image}.}
}
\value{
No value, but produces the side effect of adding a color bar to the
plot.
}
\description{
\code{add_colorbar()} adds a color bar to the plot.
}
\keyword{internal}
