% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_branch_contributions.R
\name{add_text}
\alias{add_text}
\title{Add title text to the color bar}
\usage{
add_text(dim, side = 4, line = 5, cex = 1.5, ...)
}
\arguments{
\item{dim}{An integer representing the dimension of eigenvector for plotting.}

\item{side}{See \link[graphics:mtext]{graphics::mtext}. Default to \code{4}.}

\item{line}{See \link[graphics:mtext]{graphics::mtext}. Default to \code{5}.}

\item{cex}{See \link[graphics:mtext]{graphics::mtext}. Default to \code{1.5}.}

\item{...}{Additional parameters passed to \link[graphics:mtext]{graphics::mtext}.}
}
\value{
No value, but produces the side effect of writing the title text
into the color bar.
}
\description{
\code{add_text()} adds a title text to the color bar.
}
\keyword{internal}
