% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{check_dataframe}
\alias{check_dataframe}
\title{Check if the data frame is compatible for eiger regression}
\usage{
check_dataframe(
  df_val,
  tip_labels,
  df_name = NULL,
  is_data = FALSE,
  n_tips = length(tip_labels)
)
}
\arguments{
\item{df_val}{The data frame to check.}

\item{tip_labels}{The tip labels of the phylogenetic tree for eiger
regression.}

\item{df_name}{The name of the data frame. Default to \code{NULL}.}

\item{is_data}{A boolean. \code{TRUE} if the data frame is the data argument used
for the regression, otherwise \code{FALSE}. Default to \code{FALSE}.}

\item{n_tips}{The number of tips in the phylogenetic tree. Default to
\code{length(tip_labels)}.}
}
\value{
The data frame with rows re-ordered based on \code{tip_labels} if the
data frame is valid. Raises error if:
\itemize{
\item The data frame cannot be coerced to a valid data frame,
\item The data frame does not have the same number of rows as the number of tips,
\item There are redundant labels for rows in the data frame.
Raises warning if the row names of the data frame do not match the tip
labels.
}
}
\description{
\code{check_dataframe()} checks if the data frame is compatible for eiger
regression and modifies the row orders if the data frame is valid.
}
\keyword{internal}
