% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{check_dimensions}
\alias{check_dimensions}
\title{Check if the input dimensions are integers between some lower and upper bound}
\usage{
check_dimensions(x, low, high)
}
\arguments{
\item{x}{An integer or a vector of integers for the input dimensions.}

\item{low}{An integer for the lower bound of the dimensions.}

\item{high}{An integer for the upper bound of the dimensions.}
}
\value{
Invisible returns \code{NULL} if each of the input dimensions is an
integer between some lower bound and some upper bound, otherwise raises an
error.
}
\description{
\code{check_dimensions()} checks if each of the input dimensions is an integer
between some lower bound and some upper bound.
}
\examples{
\dontrun{
check_dimensions(1, 1, 5)
check_dimensions(1:9, 1, 10)
check_dimensions(c(5, 4, 9, 8, 1, 3), 1, 10)
try(check_dimensions(-7, 1, 10))
try(check_dimensions(0, 1, 8))
try(check_dimensions(c(3, 2, 11, 4, 6), 1, 8))
try(check_dimensions(c(0.5, 2, 1, 8), 1, 8))
}

}
\keyword{internal}
