% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{check_eiger}
\alias{check_eiger}
\title{Check if the data are compatible for eiger regression}
\usage{
check_eiger(formula, data, tree, n_eigenvectors)
}
\arguments{
\item{formula}{A formula describing the model to run eiger regression. The
eigenvectors should not appear in the formula.}

\item{data}{A data frame containing the variables in the model. The data
frame should have \code{n} rows, where \code{n} is the number of tips in \code{tree}.}

\item{tree}{A phylogenetic tree as a valid object of class \code{"phylo"}.}

\item{n_eigenvectors}{The number of eigenvectors to include as additional
fixed effects in eiger regression.}
}
\value{
A data frame of \code{n} rows and \code{m} columns, where \code{n} is the number of
tips in \code{tree}, and \code{m} is the number of columns of \code{data}. The rows are
ordered according to the order of tip labels in \code{tree}. No errors raised if
all the following requirements are met:
\itemize{
\item The input \code{tree} is a valid object of class \code{"phylo"},
\item The input \code{tree} has all branch lengths defined and non-negative,
\item The input \code{tree} has all unique tip labels,
\item \code{n_eigenvectors} is less than or equal to the number of tips in \code{tree} and greater than to equal to \code{0},
\item All the variables in \code{formula} can be found in \code{data},
\item None of the variables in \code{formula} contains any \code{NA} values,
\item The response variable in \code{formula} is a vector of numeric values,
\item \code{data} is provided either as a matrix or a data frame,
\item The number of rows of \code{data} is equal to the number of tips in \code{tree},
\item The row names of \code{data} are equal to the tip labels of \code{tree},
\item All columns have names in \code{data},
\item There are no redundant column names in \code{data}.
Otherwise raises an error.
}
}
\description{
\code{check_eiger()} checks if the trait values, the phylogenetic
tree, and the number of eigenvectors to include in the regression are
compatible with each other for running eiger regression.
}
\examples{
\dontrun{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(100, 1, 1, 0, 1, complete = FALSE)[[1]]
BM <- phytools::fastBM(yule_tree, 1, nsim = 3)

BM_df <- as.data.frame(BM)
colnames(BM_df) <- c("x", "y", "z")
check_eiger(y ~ x, BM_df, yule_tree, 20)
check_eiger(y ~ x + log(z), BM_df, yule_tree, 20)
}

}
\keyword{internal}
