% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{check_list}
\alias{check_list}
\title{Check if the list can be coerced to a valid data frame}
\usage{
check_list(list_val, list_name = NULL, is_data = FALSE)
}
\arguments{
\item{list_val}{The list to check.}

\item{list_name}{The name of the list. Default to \code{NULL}.}

\item{is_data}{A boolean. \code{TRUE} if the list is the data argument used for
the regression, otherwise \code{FALSE}. Default to \code{FALSE}.}
}
\value{
The list with values of each field re-ordered and names imputed
based on the first field in the list that has names if the list is valid.
Raises error if:
\itemize{
\item The list is empty,
\item The fields in the list do not all have the same lengths,
\item The list contains fields that do not have tags,
\item The tags of the fields in the list are not all unique,
\item The list contains fields that have NA values in names,
\item The list contains fields that have redundant names.
Raises warning if:
\item There are fields in the list that have no names,
\itemize{
\item If none of the fields has names, the fields will remain having no names,
\item If only some of the fields have no names, these fields will take the names of the first field in the list that has names,
}
\item The fields in the list do have have the same set of names, in which case the fields will all will take the names of the first field in the list that has names.
}
}
\description{
\code{check_list()} checks if the list can be coerced to a valid data frame with
and modifies the names for the fields if the data frame is valid.
}
\keyword{internal}
