% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{check_names_list}
\alias{check_names_list}
\title{Check if a names list is \code{NULL} or contains \code{NA} or redundant values}
\usage{
check_names_list(
  list_val,
  null_message,
  na_top_message,
  na_bullet_prefix,
  na_bullet_suffix,
  na_extra_message,
  redundant_message
)
}
\arguments{
\item{list_val}{The names list to check.}

\item{null_message}{The error message to display if the list is \code{NULL}.}

\item{na_top_message}{A string representing the overall big issue if the list contains \code{NA} values.}

\item{na_bullet_prefix}{Part of the bullet message that goes before the listed
variant if the list contains \code{NA} values.}

\item{na_bullet_suffix}{Part of the bullet message that goes after the listed
variant if the list contains \code{NA} values.}

\item{na_extra_message}{A string stating the number of unlisted problems if the list contains \code{NA} values.}

\item{redundant_message}{The error message to display if the list contains redundant values.}
}
\value{
Invisible returns \code{NULL} if the list is not \code{NULL} and does not contain any \code{NA} or redundant
values. Otherwise raises an error.
}
\description{
\code{check_names_list()} checks if a names list is \code{NULL} or contains \code{NA} or redundant values.
}
\keyword{internal}
