% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{check_numeric_response}
\alias{check_numeric_response}
\title{Check if a response variable is numeric}
\usage{
check_numeric_response(response_val, response_name)
}
\arguments{
\item{response_val}{The response variable to check.}

\item{response_name}{The name of the response variable.}
}
\value{
Invisible returns \code{NULL} if the response variable is numeric.
Otherwise raises an error.
}
\description{
\code{check_numeric_response()} checks if a response variable is numeric.
}
\keyword{internal}
