% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{check_phylo}
\alias{check_phylo}
\title{Check if an object is a valid \code{"Phylo"} object}
\usage{
check_phylo(tree)
}
\arguments{
\item{tree}{A phylogenetic tree as an object of class \code{"phylo"}.}
}
\value{
Invisible returns \code{NULL} if the phylogenetic tree is a valid object
of class \code{"phylo"}, otherwise raises an error.
}
\description{
\code{check_phylo()} checks if an object is a valid object of class \code{"Phylo"}.
}
\examples{
\dontrun{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(10, 1, 1, 0, 1, complete = FALSE)[[1]]
check_phylo(yule_tree)
broken_tree <- list(edge = matrix(1:4, 2, 2), tip.label = letters[1:3])
try(check_phylo(broken_tree))
}

}
\keyword{internal}
