% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{check_phylo_branches}
\alias{check_phylo_branches}
\title{Check if a \code{"Phylo"} object has valid branch lengths}
\usage{
check_phylo_branches(tree)
}
\arguments{
\item{tree}{A phylogenetic tree as a valid object of class \code{"phylo"}.}
}
\value{
Invisible returns \code{NULL} if the phylogenetic tree has defined
non-negative lengths for all its branches, otherwise raises an error.
}
\description{
\code{check_phylo_branches()} checks if a \code{"Phylo"} object has defined lengths for
all its branches and if the branch lengths are all non-negative.
}
\examples{
\dontrun{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(10, 1, 1, 0, 1, complete = FALSE)[[1]]
check_phylo_branches(yule_tree)

# Input tree must have all the branch lengths defined and non-negative
broken_tree <- list(edge = matrix(c(4, 1, 4, 5, 5, 2, 5, 3), 4, 2, byrow = TRUE),
                    tip.label = letters[1:3],
                    Nnode = 2)
class(broken_tree) <- "phylo"
try(check_phylo_branches(broken_tree))
broken_tree$edge.length <- c(1, 2)
try(check_phylo_branches(broken_tree))
broken_tree$edge.length <- c(1, 2, NA, 1)
try(check_phylo_branches(broken_tree))
broken_tree$edge.length <- c(1, 2, -1, 1)
try(check_phylo_branches(broken_tree, "vcv", 2))
}

}
\keyword{internal}
