% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{check_plain}
\alias{check_plain}
\title{Check if the plain variable is compatible for eiger regression}
\usage{
check_plain(plain_val, plain, tip_labels, n_tips = length(tip_labels))
}
\arguments{
\item{plain_val}{The variable to check.}

\item{plain}{The name of the variable.}

\item{tip_labels}{The tip labels of the phylogenetic tree for eiger
regression.}

\item{n_tips}{The number of tips in the phylogenetic tree. Default to
\code{length(tip_labels)}.}
}
\value{
The plain variable with elements re-ordered based on \code{tip_labels} if
the variable is valid. Raises error if:
\itemize{
\item The variable does not have the same number of elements as the number of tips,
\item The variable contains any \code{NA} values,
\item There are redundant names for elements in the variable.
Raises warning if:
\item The names of the variable do not exist,
\item The names of the variable do not match the tip labels of the tree.
}
}
\description{
\code{check_plain()} checks if the plain variable is compatible for eiger
regression and modifies the order of the elements if the variable is valid.
}
\keyword{internal}
