% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_branch_contributions.R
\name{check_plotting_data}
\alias{check_plotting_data}
\title{Check if the data for plotting is valid}
\usage{
check_plotting_data(tree, branch_contributions, dim)
}
\arguments{
\item{tree}{A phylogenetic tree as an object of class \code{"phylo"}.}

\item{branch_contributions}{A data frame representing the contributions of
branches of the phylogenetic tree to the eigenvectors of some version of
variance-covariance matrix for the whole tree. The data frame of should
have \code{m} rows and \code{n} columns, where \code{m} is the number of branches, and \code{n}
is the number of dimensions of eigenvectors for which the branch
contributions are computed.}

\item{dim}{The dimension of eigenvector to plot.}
}
\value{
Invisible returns \code{NULL} if all the following requirements are met:
\itemize{
\item The data frame for branch contributions has the same number of rows as
the number of branches in the tree.
\item The data frame for branch contributions has the column names representing
the dimensions of eigenvectors that are less than the number of tips in the
tree.
\item The dimension for plotting is included in the data frame for branch
contributions.
Otherwise raises an error.
}
}
\description{
\code{check_plotting_data()} checks if the data frame for branch contributions is
consistent with the phylogenetic tree and if the dimension for plotting is
included in the data frame.
}
\keyword{internal}
