% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{check_redundant}
\alias{check_redundant}
\title{Check if a list contains redundant values}
\usage{
check_redundant(list_val, error_message)
}
\arguments{
\item{list_val}{The list to check.}

\item{error_message}{The error message to display if the list contains
redundant values.}
}
\value{
Invisible returns \code{NULL} if the list does not contain any redundant
values. Otherwise raises an error.
}
\description{
\code{check_redundant()} checks if a list contains any redundant values.
}
\keyword{internal}
