% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_branch_contributions.R
\name{check_valid_color}
\alias{check_valid_color}
\title{Check if an input color is valid}
\usage{
check_valid_color(col)
}
\arguments{
\item{col}{A string representing color.}
}
\value{
\code{TRUE} if the color is valid and \code{FALSE} otherwise.
}
\description{
\code{check_valid_color()} checks if an input color is valid.
}
\examples{
\dontrun{
check_valid_color("blue")
check_valid_color("#BBCCEE")
check_valid_color("apple")
}

}
\keyword{internal}
