% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{compute_branch_contributions}
\alias{compute_branch_contributions}
\title{Compute branch contributions to eigenvectors of variance-covariance matrices}
\usage{
compute_branch_contributions(
  tree,
  cov_matrix = c("dcvcv", "vcv", "egrm"),
  dims
)
}
\arguments{
\item{tree}{A phylogenetic tree as an object of class \code{"phylo"} with all the
branch lengths defined and non-negative.}

\item{cov_matrix}{A string representing the way to calculate the
variance-covariance matrices. The options are \code{"dcvcv"} (double-centered
phylogenetic variance-covariance matrix), \code{"vcv"} (non-centered
phylogenetic variance-covariance matrix), and \code{"egrm"} (expected genetic
relatedness matrix). Default to \code{"dcvcv"}.}

\item{dims}{An integer or a vector of integers for the dimensions of
eigenvectors to compute the branch contributions.}
}
\value{
A data frame of \code{m} rows and \code{n} columns, where \code{m} is the number of
branches, and \code{n} is the number of dimensions of eigenvectors to compute
the branch contributions. The \code{i}-th row corresponds to the contributions
of the \code{i}-th branch to the eigenvectors, and the \code{j}-th column has name
\code{"dim_k"}, where \code{k} is the . The last column refers to where \code{k} is the
\code{j}-th dimension in \code{dims} to compute the branch contributions. Raises
error if the input tree is not a valid tree of class \code{"phylo"} or if not
all the branch lengths are defined or if some branch lengths are negative.
Also raises error if the input dimensions contain either non-integers or
integers less than \code{1} or greater than the number of tips in the tree.
}
\description{
\code{compute_branch_contributions()} computes the contributions of each branch of
a phylogenetic tree to the eigenvectors of some version of
variance-covariance matrix for the whole tree.
}
\examples{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(10, 1, 1, 0, 1, complete = FALSE)[[1]]
compute_branch_contributions(yule_tree, "vcv", 1)
compute_branch_contributions(yule_tree, "egrm", 3)
compute_branch_contributions(yule_tree, "dcvcv", 1:5)
compute_branch_contributions(yule_tree, dims = c(2, 4, 6, 7))

# Input tree must be a valid object of class "phylo"
broken_tree <- list(edge = matrix(1:4, 2, 2), tip.label = letters[1:3])
try(compute_branch_contributions(broken_tree, "vcv", 2))

# Input tree must have all the branch lengths defined and non-negative
broken_tree <- list(edge = matrix(c(4, 1, 4, 5, 5, 2, 5, 3), 4, 2, byrow = TRUE),
                    tip.label = letters[1:3],
                    Nnode = 2)
class(broken_tree) <- "phylo"
try(compute_branch_contributions(broken_tree, "vcv", 2))
broken_tree$edge.length <- c(1, 2)
try(compute_branch_contributions(broken_tree, "vcv", 2))
broken_tree$edge.length <- c(1, 2, NA, 1)
try(compute_branch_contributions(broken_tree, "vcv", 2))
broken_tree$edge.length <- c(1, 2, -1, 1)
try(compute_branch_contributions(broken_tree, "vcv", 2))

# Input dimensions must all be integers from 1 to the number of tips in the tree
try(compute_branch_contributions(yule_tree, "vcv", 11))
try(compute_branch_contributions(yule_tree, "dcvcv", c(8, 9, 2, 14, 1)))

}
