% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{compute_branch_egrm}
\alias{compute_branch_egrm}
\title{Construct the expected genetic relatedness matrix for a branch}
\usage{
compute_branch_egrm(x)
}
\arguments{
\item{x}{A binary vector for the descendancy relationships between the tips
and the branch. An entry is \code{1} if the corresponding tip is a descendant of
the branch cocnerned and \code{0} otherwise.}
}
\value{
The expected genetic variance-covariance matrix for the branch. An
entry in the matrix is positive if the two tips corresponding to the row
and the column are both descendants or non-descendants of the branch, and
negative if one tip is a descendant and the other is not. The exact values
depend on the partition of all the tips into descendants and
non-descendants.
}
\description{
\code{compute_branch_egrm()} computes the expected genetic relatedness matrix for
a branch of a phylogenetic tree.
}
\examples{
\dontrun{
compute_branch_egrm(c(1, 0, 0, 1, 1))
}

}
\keyword{internal}
