% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{compute_branch_vcv}
\alias{compute_branch_vcv}
\title{Construct the variance-covariance matrix for a branch}
\usage{
compute_branch_vcv(x)
}
\arguments{
\item{x}{A binary vector for the descendancy relationships between the tips
and the branch. An entry is \code{1} if the corresponding tip is a descendant of
the branch cocnerned and \code{0} otherwise.}
}
\value{
A binary variance-covariance matrix for the branch. An entry in the
matrix is \code{1} if the two tips corresponding to the row and the column are
both descendants of the branch, and \code{0} otherwise.
}
\description{
\code{compute_branch_vcv()} computes the variance-covariance matrix for a branch
of a phylogenetic tree. An entry in the matrix is \code{1} if the two tips
corresponding to the row and the column are both descendants of the branch,
and \code{0} otherwise.
}
\examples{
\dontrun{
compute_branch_vcv(c(1, 0, 0, 1, 1))
}

}
\keyword{internal}
