% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{compute_cov_matrices}
\alias{compute_cov_matrices}
\title{Construct variance-covariance matrices for a tree and each of its branches}
\usage{
compute_cov_matrices(tree, cov_matrix = c("dcvcv", "vcv", "egrm"))
}
\arguments{
\item{tree}{A phylogenetic tree as an object of class \code{"phylo"} with all the
branch lengths defined and non-negative.}

\item{cov_matrix}{A string representing the way to calculate the
variance-covariance matrices. The options are \code{"dcvcv"} (double-centered
phylogenetic variance-covariance matrix), \code{"vcv"} (non-centered
phylogenetic variance-covariance matrix), and \code{"egrm"} (expected genetic
relatedness matrix). Default to \code{"dcvcv"}.}
}
\value{
A list of two values.

The first value \code{cov_matrices} is an \code{n^2} by \code{(m + 1)} dataframe, where
\code{n} is the number of tips in the phylogenetic tree and \code{m} is the number of
branches. The first \code{m} columns refer to the variance-covariance matrices
for the corresponding branches in the tree. The \code{k}-th of the first \code{m}
columns has name \code{"dcvcv_k"}, \code{"vcv_k"}, or \code{"egrm_k"} depending on the way
the variance-covariance matrices are calculated. The last column refers to
the variance-covariance matrix for the whole tree and has name
\code{"dcvcv_tot"}, \code{"vcv_tot"}, or \code{"egrm_tot"}. The matrices are collapsed
into vectors for easier visualization and downstream analyses.

The second value \code{cov_matrix_tree} is the variance-covariance matrix for
the whole phylogenetic tree.

Raises error if the input is not a valid tree of class \code{"phylo"} or if not
all the branch lengths are defined or if some branch lengths are negative.
}
\description{
\code{compute_cov_matrices()} computes some version of variance-covariance matrix
for a phylogenetic tree as well as for each of its branches.
}
\examples{
\dontrun{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(10, 1, 1, 0, 1, complete = FALSE)[[1]]
compute_cov_matrices(yule_tree)
compute_cov_matrices(yule_tree, cov_matrix = "egrm")

# Input tree must be a valid object of class "phylo"
broken_tree <- list(edge = matrix(1:4, 2, 2), tip.label = letters[1:3])
try(compute_cov_matrices(broken_tree))

# Input tree must have all the branch lengths defined and non-negative
broken_tree <- list(edge = matrix(c(4, 1, 4, 5, 5, 2, 5, 3), 4, 2, byrow = TRUE),
                    tip.label = letters[1:3],
                    Nnode = 2)
class(broken_tree) <- "phylo"
try(compute_cov_matrices(broken_tree))
broken_tree$edge.length <- c(1, 2)
try(compute_cov_matrices(broken_tree))
broken_tree$edge.length <- c(1, 2, NA, 1)
try(compute_cov_matrices(broken_tree))
broken_tree$edge.length <- c(1, 2, -1, 1)
try(compute_cov_matrices(broken_tree))
}

}
\keyword{internal}
