% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{compute_cov_matrix}
\alias{compute_cov_matrix}
\title{Construct the variance-covariance matrix for a phylogenetic tree}
\usage{
compute_cov_matrix(tree, cov_matrix = c("dcvcv", "vcv", "egrm"))
}
\arguments{
\item{tree}{A phylogenetic tree as an object of class \code{"phylo"} with all the
branch lengths defined and non-negative.}

\item{cov_matrix}{A string representing the way to calculate the
variance-covariance matrices. The options are \code{"dcvcv"} (double-centered
phylogenetic variance-covariance matrix), \code{"vcv"} (non-centered
phylogenetic variance-covariance matrix), and \code{"egrm"} (expected genetic
relatedness matrix). Default to \code{"dcvcv"}.}
}
\value{
A matrix representing the variance-covariance matrix for the whole
phylogenetic tree. Raises error if the input is not a valid tree of class
\code{"phylo"} or if not all the branch lengths are defined or if some branch
lengths are negative.
}
\description{
\code{compute_cov_matrix()} computes some version of variance-covariance matrix
for a phylogenetic tree.
}
\examples{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(10, 1, 1, 0, 1, complete = FALSE)[[1]]
compute_cov_matrix(yule_tree)
compute_cov_matrix(yule_tree, cov_matrix = "egrm")

# Input tree must be a valid object of class "phylo"
broken_tree <- list(edge = matrix(1:4, 2, 2), tip.label = letters[1:3])
try(compute_cov_matrix(broken_tree))

# Input tree must have all the branch lengths defined and non-negative
broken_tree <- list(edge = matrix(c(4, 1, 4, 5, 5, 2, 5, 3), 4, 2, byrow = TRUE),
                    tip.label = letters[1:3],
                    Nnode = 2)
class(broken_tree) <- "phylo"
try(compute_cov_matrix(broken_tree))
broken_tree$edge.length <- c(1, 2)
try(compute_cov_matrix(broken_tree))
broken_tree$edge.length <- c(1, 2, NA, 1)
try(compute_cov_matrix(broken_tree))
broken_tree$edge.length <- c(1, 2, -1, 1)
try(compute_cov_matrix(broken_tree))

}
