% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{compute_descendancy_matrix}
\alias{compute_descendancy_matrix}
\title{Calculate the descendancy matrix for branches of a tree}
\usage{
compute_descendancy_matrix(tree)
}
\arguments{
\item{tree}{A phylogenetic tree as an object of class \code{"phylo"}.}
}
\value{
A binary descendancy matrix with each row representing a branch of
the tree and each column a tip of the tree. An entry of the matrix is \code{1}
if the corresponding tip (column) is a descendant of the corresponding
branch (row) and \code{0} otherwise. Raises error if the input is not a valid
tree of class \code{"phylo"}.
}
\description{
\code{compute_descendancy_matrix()} computes a binary matrix for descendancy
relationships between branches and tips of a tree, with an entry being \code{1} if
the corresponding tip (column) is a descendant of the corresponding branch
(row) and \code{0} otherwise.
}
\examples{
\dontrun{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(10, 1, 1, 0, 1, complete = FALSE)[[1]]
compute_descendancy_matrix(yule_tree)

# Input must be a valid object of class "phylo"
broken_tree <- list(edge = matrix(1:4, 2, 2), tip.label = letters[1:3])
try(compute_descendancy_matrix(broken_tree))
}

}
\keyword{internal}
