% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_contributions.R
\name{create_bullet_error}
\alias{create_bullet_error}
\title{Create and throw an error message that lists the individual problems}
\usage{
create_bullet_error(
  invalid_items,
  top_message,
  bullet_prefix,
  bullet_suffix,
  extra_message
)
}
\arguments{
\item{invalid_items}{A vector of invalid items that lead to problems.}

\item{top_message}{A string representing the overall big issue.}

\item{bullet_prefix}{Part of the bullet message that goes before the listed
variant.}

\item{bullet_suffix}{Part of the bullet message that goes after the listed
variant.}

\item{extra_message}{A string stating the number of unlisted problems.}
}
\value{
No return value, always throws an error. The first line of the error
message states the overal big issue. The next few lines are bullets
describing the individual problems in order. If there are more than five
problems, only the first five problems are listed, and an extra line shows
the number of extra problems not listed in the bullets.
}
\description{
\code{create_bullet_error()} creates and throws an error message that states the
overall error, lists the individual problems as bullets, and states the
number of extra problems if there are more than five problems.
}
\keyword{internal}
