% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_branch_contributions.R
\name{create_color_palette}
\alias{create_color_palette}
\title{Create a color palette from a color-generating function or a set of colors}
\usage{
create_color_palette(color_scheme = NULL, n_colors = 100)
}
\arguments{
\item{color_scheme}{A string representing a color-generating function or
vector of strings representing the set of colors to interpolate to create
the new palette. Default to \code{NULL}.}

\item{n_colors}{An integer representing the number of colors in the color
palette. Default to \code{100}.}
}
\value{
A vector of strings representing the set of colors in the new color
palette. Raises error if the name of the color-generating function is not
valid or if the list of colors provided are not valid.
}
\description{
\code{create_color_palette()} creates a new color palette by extracting colors
from from a color-generating function or interpolating a set of given colors.
}
\examples{
\dontrun{
create_color_palette(c("red", "blue"), 30)
create_color_palette(c("#BBCCEE", "#CCEEFF", "#CCDDAA", "#EEEEBB", "#FFCCCC"), 50)
create_color_palette("Viridis", 80)
try(create_color_palette("red", 30))
try(create_color_palette(c("apple", "blueberry"), 30))
try(create_color_palette(c("red", "blue"), 3.14))
try(create_color_palette("CoolColors", 80))
}

}
\keyword{internal}
