% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_branch_contributions.R
\name{get_default_colors}
\alias{get_default_colors}
\title{Create the default list of colors}
\usage{
get_default_colors()
}
\value{
A vector of strings representing colors from dark blue to light blue
to orange to dark red.
}
\description{
\code{get_default_colors()} creates the default list of colors for downstream
plotting.
}
\keyword{internal}
