% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_branch_contributions.R
\name{map_colors}
\alias{map_colors}
\title{Map a vector of values to colors in the color palette}
\usage{
map_colors(x, color_palette)
}
\arguments{
\item{x}{A vector of values to be mapped to colors.}

\item{color_palette}{A list of strings representing the color palette for
mapping.}
}
\value{
A list of two values.

The first value \code{color_breaks} is a vector representing the boundaries of
value intervals so that values falling in the same interval would be
assigned the same color.

The second value \code{mapped_colors} is a vector of strings representing the
colors each value is mapped to.
}
\description{
\code{map_colors()} maps a vector of values to colors in the color palette. The
mapping is performed in a way such that the values close to \code{0} would be
mapped to colors in the middle of the color palette. Depending on the
distribution of the values, the first or the last color in the color palette
might not necessarily have a value being mapped to.
}
\examples{
\dontrun{
map_colors(-8:8, c("blue", "red"))
x <- c(-3.5, 9.8, 0.3, -1.1, -4.6, 2.1)
color_palette <- c("#BBCCEE", "#CCEEFF", "#CCDDAA", "#EEEEBB", "#FFCCCC")
map_colors(x, color_palette)
}

}
\keyword{internal}
