% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse the formula into names for data frames, fields, and plain variables}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{A formula to parse.}
}
\value{
A list containing three vectors: \code{dfs} for the names of data frames
(before dollar signs), \code{fields} for the names of fields (after dollar
signs), and \code{plains} for the names of plain variables (no dollar signs).
}
\description{
\code{parse_formula()} gets the names for data frames, fields, and plain variables
from a formula.
}
\examples{
\dontrun{
parse_formula(y ~ x)
parse_formula(y ~ x + 1)
parse_formula(y ~ x + log(z))
parse_formula(y ~ df1$x + df2$a + b)
parse_formula(y ~ df1$x + df2$x + x)
}

}
\keyword{internal}
