% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_branch_contributions.R
\name{plot_branch_contributions}
\alias{plot_branch_contributions}
\title{Plot branch contributions to an eigenvector of variance-covariance matrices}
\usage{
plot_branch_contributions(
  tree,
  branch_contributions = NULL,
  cov_matrix = c("dcvcv", "vcv", "egrm"),
  dim,
  text_args = list(),
  axis_args = list(),
  colorbar_args = list(),
  color_scheme = NULL,
  n_colors = 100,
  edge.width = 5,
  show.tip.label = FALSE,
  ...
)
}
\arguments{
\item{tree}{A phylogenetic tree as an object of class \code{"phylo"}.}

\item{branch_contributions}{A data frame representing the contributions of
branches of the phylogenetic tree to the eigenvectors of some version of
variance-covariance matrix for the whole tree. The data frame of should
have \code{m} rows and \code{n} columns, where \code{m} is the number of branches, and \code{n}
is the number of dimensions of eigenvectors for which the branch
contributions are computed. Default to \code{NULL}, in which case
\code{branch_contributions} will be calculated using \code{cov_matrix}.}

\item{cov_matrix}{A string representing the version of variance-covariance
matrix used to obtain the data frame \code{branch_contributions}. The options
are \code{"dcvcv"} (double-centered phylogenetic variance-covariance matrix),
\code{"vcv"} (non-centered phylogenetic variance-covariance matrix), and
\code{"egrm"} (expected genetic relatedness matrix). If \code{branch_contributions}
is \code{NULL}, \code{cov_matrix} is used to calculate \code{branch_contributions}.
Default to \code{"dcvcv"}.}

\item{dim}{The dimension of eigenvector to plot.}

\item{text_args}{A named list of arguments passed to an internal helper that
adds title text to the color bar. Elements include:
\describe{
\item{side}{See \link[graphics:mtext]{graphics::mtext}. Default to \code{4}.}
\item{line}{See \link[graphics:mtext]{graphics::mtext}. Default to \code{5}.}
\item{cex}{See \link[graphics:mtext]{graphics::mtext}. Default to \code{1.5}.}
\item{...}{Additional parameters passed to \link[graphics:mtext]{graphics::mtext}.}
}}

\item{axis_args}{A named list of arguments passed to an internal helper that
adds axis to the color bar.
\describe{
\item{side}{See \link[graphics:axis]{graphics::axis}. Default to \code{4}.}
\item{cex.axis}{See \link[graphics:axis]{graphics::axis}. Default to \code{1.5}.}
\item{...}{Additional parameters passed to \link[graphics:axis]{graphics::axis}.}
}}

\item{colorbar_args}{Parameters passed to \link[graphics:image]{graphics::image} for making the
color bar.}

\item{color_scheme}{A string representing a color-generating function or
vector of strings representing the set of colors to interpolate to create
the new palette. Default to \code{NULL}.}

\item{n_colors}{An integer representing the number of colors in the color
palette. Default to \code{100}.}

\item{edge.width}{See \link[ape:plot.phylo]{ape::plot.phylo}. Default to \code{5}.}

\item{show.tip.label}{See \link[ape:plot.phylo]{ape::plot.phylo}. Default to \code{FALSE}.}

\item{...}{Other parameters passed to \link[ape:plot.phylo]{ape::plot.phylo}.}
}
\value{
No value, but produces the side effect of plotting the contributions
of branches of a phylogenetic tree to the eigenvector of dimension \code{dim}
using some version of variance-covariance matrix for the whole tree. A
color bar is also included with the plot.
}
\description{
\code{plot_branch_contributions()} plots the contributions of branches of a
phylogenetic tree to some eigenvector using some version of
variance-covariance matrix for the whole tree.
}
\examples{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(20, 1, 1, 0, 1, complete = FALSE)[[1]]
df <- compute_branch_contributions(yule_tree, "egrm", 1:10)
plot_branch_contributions(yule_tree, df, "egrm", 3)
plot_branch_contributions(yule_tree, df, "egrm", 5, n_colors = 50)
plot_branch_contributions(yule_tree, cov_matrix = "dcvcv", dim = 5, n_colors = 80)

}
