% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{prepare_eiger}
\alias{prepare_eiger}
\title{Prepare the data frame for running eiger regression}
\usage{
prepare_eiger(
  formula = NULL,
  data = NULL,
  tree,
  n_eigenvectors,
  x = NULL,
  y = NULL,
  dc = FALSE
)
}
\arguments{
\item{formula}{A formula describing the model to run eiger regression. If
supplied, \code{x} and \code{y} are ignored. The eigenvectors should not appear in
the formula. Default to \code{NULL}.}

\item{data}{A data frame containing the variables in the model. The data
frame should have \code{n} rows, where \code{n} is the number of tips in \code{tree}. If
not supplied or if variables are not found in \code{data}, the variables are
taken from the current environment. Default to \code{NULL}.}

\item{tree}{A phylogenetic tree as a valid object of class \code{"phylo"}.}

\item{n_eigenvectors}{The number of eigenvectors to include as additional
fixed effects in eiger regression.}

\item{x}{A vector representing the predictor variables of the \code{n} tips in the
tree. Used only when \code{formula} is missing. Default to \code{NULL}.}

\item{y}{A vector representing the outcome variables of the \code{n} tips in the
tree. The values must be numeric. Used only when \code{formula} is missing.
Default to \code{NULL}.}

\item{dc}{A boolean value. \code{TRUE} if the eigenvectors need to be calculated
from double-centered variance-covariance matrix and \code{FALSE} otherwise.
Default to \code{FALSE}.}
}
\value{
A data frame of \code{n} rows, where \code{n} is the number of tips in \code{tree}.
If \code{x} and \code{y} are used, the output data frame contains \code{(n + 2)} columns:
the first two columns have names "X" and "Y", and the next \code{n} columns
represent the \code{n} eigenvectors of the variance-covariance matrix; if
\code{formula} is used, the output data frame contains \code{(m + k + n)} columns,
where the first \code{m} columns are the columns in \code{data}, the next \code{k} columns
are the additional variables defined in \code{formula} but not found in \code{data},
and the last \code{n} columns represent the \code{n} eigenvectors of the
variance-covariance matrix. The rows are ordered according to the order of
tip labels in \code{tree}. If \code{data} or any other data frame used is provided in
terms of a list, the list has values of each field re-ordered and names
imputed based on the first field in the list that has names before being
coerced to data frame. No errors raised if all the following requirements
are met:
\itemize{
\item The input \code{tree} is a valid object of class \code{"phylo"},
\item The input \code{tree} has all branch lengths defined and non-negative,
\item The input \code{tree} has all unique tip labels,
\item \code{n_eigenvectors} is less than or equal to the number of tips in \code{tree} and greater than to equal to \code{0},
\item If \code{x} and \code{y} are used:
\itemize{
\item \code{y} is a vector of numeric values,
\item The lengths of \code{x} and \code{y} are equal to the number of tips in \code{tree},
\item \code{x} and \code{y} do not contain any \code{NA} values,
\item There are no redundant names in \code{x} and \code{y} if names exist,
}
\item If \code{formula} is used:
\itemize{
\item All the variables in \code{formula} can be found either in \code{data} or in the current environment,
\item The lengths of all the variables in \code{formula} are equal to the number of tips in \code{tree},
\item None of the variables in \code{formula} contains any \code{NA} values,
\item The response variable in \code{formula} is a vector of numeric values,
\item If a variable in \code{formula} is in the current environment:
\itemize{
\item There are no redundant names in the variable if names exist,
}
\item If a variable in \code{formula} is a field of some other data frame:
\itemize{
\item The data frame exists in the current environment and the variable is a column of that data frame,
\item The number of rows of the data frame is equal to the number of tips in \code{tree},
\item All columns have names in the data frame,
\item There are no redundant column names in the data frame,
}
}
\item If \code{data} is supplied:
\itemize{
\item \code{data} can be coerced to a valid data frame,
\item The number of rows of \code{data} is equal to the number of tips in \code{tree},
\item All columns have names in \code{data},
\item There are no redundant column names in \code{data},
}
\item When \code{data} or other data frames used are provided in terms of a list:
\itemize{
\item The list is not empty,
\item The fields in the list all have the same lengths,
\item All the fields in the list has tags,
\item The tags of the fields in the list are all unique,
\item The list does not contain fields that have NA values in names,
\item The list does not contain fields that have redundant names.
Otherwise raises an error.
}
}

Raises a warning if:
\itemize{
\item \code{x} and \code{y} are used and either don't have names or have names not equal to the tip labels of \code{tree},
\item The row names of \code{data} are not equal to the tip labels of \code{tree},
\item A variable in \code{formula} is a vector found in the current environment and either don't have names or have names not equal to the tip labels of \code{tree},
\item A variable in \code{formula} is column of a data frame found in the current environment and the row names of the data frame are not equal to the tip labels of \code{tree}.
\item When \code{data} or other data frames used are provided in terms of a list:
\itemize{
\item There are fields in the list that have no names,
\itemize{
\item If none of the fields has names, the fields will remain having no names,
\item If only some of the fields have no names, these fields will take the names of the first field in the list that has names,
}
\item The fields in the list do have have the same set of names, in which case the fields will all will take the names of the first field in the list that has names.
}
}
}
\description{
\code{prepare_eiger()} constructs a data frame that is essential for
running eiger regression. The data frame includes all the original data
provided for the regression as well as the eigenvectors of the
variance-covariance matrix.

Users can either provide:
\itemize{
\item The predictor and outcome variables \code{x} and \code{y}.
\item \code{formula} and \code{data}.
}
}
\examples{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(100, 1, 1, 0, 1, complete = FALSE)[[1]]
BM <- phytools::fastBM(yule_tree, 1, nsim = 3)
x <- BM[, 1]
y <- BM[, 2]
z <- BM[, 3]

BM_df <- as.data.frame(BM)
colnames(BM_df) <- c("x", "y", "z")
prepare_eiger(y ~ x, BM_df, yule_tree, 20)
prepare_eiger(y ~ x + log(z), BM_df, yule_tree, 20, dc = TRUE)

group <- c(rep("A", 40), rep("B", 60))
suppressWarnings(prepare_eiger(y ~ x + group, BM_df, yule_tree, 20))
suppressWarnings(prepare_eiger(BM_df$y ~ BM_df$x + group, tree = yule_tree, n_eigenvectors = 20))

suppressWarnings(prepare_eiger(tree = yule_tree, n_eigenvectors = 20, x = x, y = y))

}
