% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{run_eiger}
\alias{run_eiger}
\title{Run eiger regression}
\usage{
run_eiger(
  formula,
  data,
  tree,
  n_eigenvectors,
  dc = FALSE,
  prepared = FALSE,
  intercept = TRUE,
  ...
)
}
\arguments{
\item{formula}{A formula describing the model to run eiger regression. The
eigenvectors should not appear in the formula.}

\item{data}{A data frame of \code{n} rows, where \code{n} is the number of tips in
\code{tree}. If \code{prepared} is \code{FALSE}, the data frame should contain the
variables in the model. If \code{prepared}
is \code{TRUE}, the data frame should be the output data frame from
\link{prepare_eiger}.}

\item{tree}{A phylogenetic tree as a valid object of class \code{"phylo"}.}

\item{n_eigenvectors}{The number of eigenvectors to include as additional
fixed effects in eiger regression.}

\item{dc}{A boolean value. \code{TRUE} if the eigenvectors need to be calculated
from double-centered variance-covariance matrix and \code{FALSE} otherwise.
Default to \code{FALSE}.}

\item{prepared}{A boolean. \code{TRUE} if the output data frame from
\link{prepare_eiger} is used, otherwise \code{FALSE}. Default to \code{FALSE}.}

\item{intercept}{A boolean. \code{TRUE} if the intercept is included in the
regression, otherwise \code{FALSE}. Default to \code{TRUE}.}

\item{...}{Other parameters passed to \link[phylolm:phylolm]{phylolm::phylolm}.}
}
\value{
The results of eiger regession, which is the same as the return
values of \link[phylolm:phylolm]{phylolm::phylolm}. Raises error if:
\itemize{
\item The input \code{tree} is not a valid object of class \code{"phylo"},
\item The input \code{tree} does not have all branch lengths defined or have some of them being negative,
\item The input \code{tree} does not have all unique tip labels,
\item \code{n_eigenvectors} is greater than the number of tips in \code{tree} or less than 0,
\item When \code{prepared} is \code{FALSE}:
\itemize{
\item Some variable in \code{formula} can't be found in \code{data},
\item Some variable in \code{formula} contains \code{NA} values,
\item The response variable in \code{formula} is a not vector of numeric values,
\item \code{data} is not provided as a matrix or a data frame,
\item The number of rows of \code{data} is not equal to the number of tips in \code{tree},
\item The row names of \code{data} are equal to the tip labels of \code{tree},
\item There are columns with no names in \code{data},
\item There are redundant column names in \code{data},
}
\item When \code{prepared} is \code{TRUE}:
\itemize{
\item \code{data} does not exist,
\item \code{data} does not contain all the variables in \code{formula},
\item \code{data} does not have all the eigenvectors of the variance-covariance matrix,
}
\item \code{n_eigenvectors} is too large (even though it's valid),
\item Other problems when running \link[phylolm:phylolm]{phylolm::phylolm}.
}
}
\description{
\code{run_eiger()} runs the eiger regression, which includes
eigenvectors of the variance-covariance matrix of a phylogenetic tree as
fixed effects in \code{phylolm}.
\itemize{
\item If users have not called \link{prepare_eiger} for the regression, to build the regression model, users should provide \code{formula} and \code{data}.
\item If users have called \link{prepare_eiger} for the regression, users should provide both the output data frame from \link{prepare_eiger} as \code{data} argument and the original formula as \code{formula}.
}
}
\examples{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(100, 1, 1, 0, 1, complete = FALSE)[[1]]
BM <- phytools::fastBM(yule_tree, 1, nsim = 3)
x <- BM[, 1]
y <- BM[, 2]
z <- BM[, 3]

BM_df <- as.data.frame(BM)
colnames(BM_df) <- c("x", "y", "z")
run_eiger(y ~ x, BM_df, yule_tree, 20)
run_eiger(y ~ x + log(z), BM_df, yule_tree, 20, dc = TRUE, intercept = FALSE)

group <- c(rep("A", 40), rep("B", 60))
try(run_eiger(y ~ x + group, BM_df, yule_tree, 20))

}
