% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{check_eiger}
\alias{check_eiger}
\title{Check if the data are compatible for running eiger regression.}
\usage{
check_eiger(x, y, tree, n_eigenvectors)
}
\arguments{
\item{x}{A vector of numeric values representing the predictor trait values
of the \code{n} tips in the tree.}

\item{y}{A vector of numeric values representing the outcome trait values of
the \code{n} tips in the tree.}

\item{tree}{A phylogenetic tree as a valid object of class \code{"phylo"}.}

\item{n_eigenvectors}{The number of eigenvectors to include as fixed effects
in eiger regression.}
}
\value{
Invisible returns \code{NULL} if all the following requirements are met:
\itemize{
\item The input \code{tree} is a valid object of class \code{"phylo"},
\item The input \code{tree} has all branch lengths defined and non-negative,
\item The input \code{tree} has all unique tip labels,
\item \code{n_eigenvectors} is less than or equal to the number of tips in \code{tree} and greater than to equal to \code{0},
\item \code{x} and \code{y} are vectors of numeric values,
\item The lengths of \code{x} and \code{y} are equal to the number of tips in \code{tree}.
Otherwise raises an error.
}
}
\description{
\code{check_eiger()} checks if the trait values, the phylogenetic tree, and the
number of eigenvectors to include in the regression are compatible with each
other.
}
\keyword{internal}
