% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eiger.R
\name{run_eiger}
\alias{run_eiger}
\title{Run eiger regression}
\usage{
run_eiger(
  x,
  y,
  tree,
  n_eigenvectors,
  dc = FALSE,
  df = NULL,
  intercept = TRUE,
  ...
)
}
\arguments{
\item{x}{A vector of numeric values representing the predictor trait values
of the \code{n} tips in the tree.}

\item{y}{A vector of numeric values representing the outcome trait values of
the \code{n} tips in the tree.}

\item{tree}{A phylogenetic tree as a valid object of class \code{"phylo"}.}

\item{n_eigenvectors}{The number of eigenvectors to include as fixed effects
in eiger regression.}

\item{dc}{A boolean value. \code{TRUE} if the eigenvectors need to be calculated
from double-centered variance-covariance matrix and \code{FALSE} otherwise.
Default to \code{FALSE}.}

\item{df}{A data frame of \code{n} rows and \code{(n + 2)} columns, where \code{n} is the
number of tips in the tree. The first two columns have names \code{X} and \code{Y},
representing the trait values. The next \code{n} columns represent the \code{n}
eigenvectors of the variance-covariance matrix. Default to \code{NULL}, in which
case \code{df} will be computed from the other parameters.}

\item{intercept}{A boolean. \code{TRUE} if the intercept is included in the
regression, otherwise \code{FALSE}. Default to \code{TRUE}.}

\item{...}{Other parameters passed to \link[phylolm:phylolm]{phylolm::phylolm}.}
}
\value{
The results of eiger regession, which is the same as the return
values of \link[phylolm:phylolm]{phylolm::phylolm}. Raises error if:
\itemize{
\item The input \code{tree} is not a valid object of class \code{"phylo"},
\item The input \code{tree} does not have all branch lengths defined or have some of them being negative,
\item The input \code{tree} does not have all unique tip labels,
\item \code{n_eigenvectors} is greater than the number of tips in \code{tree},
\item \code{x} and \code{y} are not vectors of numeric values,
\item The lengths of \code{x} and \code{y} are not equal to the number of tips in \code{tree},
\item \code{n_eigenvectors} is too large (even though it's valid),
\item Other problems when running \link[phylolm:phylolm]{phylolm::phylolm}.
}
}
\description{
\code{run_eiger()} runs the eiger regression, which includes eigenvectors of the
variance-covariance matrix of a phylogenetic tree as fixed effects in
\code{phylolm}.
}
\examples{
set.seed(42)
yule_tree <- TreeSim::sim.bd.taxa(100, 1, 1, 0, 1, complete = FALSE)[[1]]
BM <- phytools::fastBM(yule_tree, 1, nsim = 2)
x <- BM[, 1]
y <- BM[, 2]
run_eiger(x, y, yule_tree, n_eigenvectors = 20)
run_eiger(x, y, yule_tree, 30, dc = TRUE, intercept = FALSE)

}
